Add Integral SIZE to adapter
