+++ projects/integral/abis/sizeGetReserves.json
@@ -1,0 +1,20 @@
+{
+  "constant": true,
+  "inputs": [],
+  "name": "getReserves",
+  "outputs": [
+      {
+          "internalType": "uint112",
+          "name": "reserve0",
+          "type": "uint112"
+      },
+      {
+          "internalType": "uint112",
+          "name": "reserve1",
+          "type": "uint112"
+      }
+  ],
+  "payable": false,
+  "stateMutability": "view",
+  "type": "function"
+}
